+++ include/grpcpp/impl/codegen/channel_interface.h
@@ -25,12 +25,12 @@
 #include <grpcpp/impl/codegen/time.h>
 
 namespace grpc_impl {
+class ClientContext;
 class CompletionQueue;
 }
 
 namespace grpc {
 class ChannelInterface;
-class ClientContext;
 
 template <class R>
 class ClientReader;
@@ -129,7 +129,7 @@ class ChannelInterface {
   friend class ::grpc::internal::RpcMethod;
   friend class ::grpc::internal::InterceptedChannel;
   virtual internal::Call CreateCall(const internal::RpcMethod& method,
-                                    ClientContext* context,
+                                    ::grpc_impl::ClientContext* context,
                                     ::grpc_impl::CompletionQueue* cq) = 0;
   virtual void PerformOpsOnCall(internal::CallOpSetInterface* ops,
                                 internal::Call* call) = 0;
@@ -149,7 +149,7 @@ class ChannelInterface {
   // method and adding a new pure method to an interface would be a breaking
   // change (even though this is private and non-API)
   virtual internal::Call CreateCallInternal(const internal::RpcMethod& method,
-                                            ClientContext* context,
+                                            ::grpc_impl::ClientContext* context,
                                             ::grpc_impl::CompletionQueue* cq,
                                             size_t interceptor_pos) {
     return internal::Call();
