+++ include/grpcpp/impl/codegen/client_callback.h
@@ -31,11 +31,10 @@
 
 namespace grpc_impl {
 class Channel;
+class ClientContext;
 }
 
 namespace grpc {
-
-class ClientContext;
 
 namespace internal {
 class RpcMethod;
